refactor: consolidate finding-format/classifier/fingerprint into finding-factory
Closes #1336

- Merge finding-format.mjs (137 lines), finding-classifier.mjs (97 lines),
  and finding-fingerprint.mjs (29 lines) into src/lib/finding-factory.mjs
- Update all importers in src/ and tests/ to use finding-factory.mjs
- Delete the three original files

No behavior change; all 1557 tests pass.
