+++ src/lib/reviewer-orchestrator.mjs
@@ -1,5 +1,5 @@
 import { generateReview } from './review-engine.mjs';
-import { classifyFindings } from './finding-classifier.mjs';
+import { classifyFindings } from './finding-factory.mjs';
 import { renderDiffText } from './diff-optimizer.mjs';
 import { synthesizeTeamLeadReport } from './team-lead-synthesizer.mjs';
